day_11 complete

+++ day_11/src/main/java/com/ericgha/monkey/TurnResult.java
@@ -2,6 +2,6 @@ package com.ericgha.monkey;
 
 import java.util.Queue;
 
-record TurnResult(Queue<Long> trueItems, Queue<Long> falseItems) {
+record TurnResult(Queue<Integer> trueItems, Queue<Integer> falseItems) {
 
 }
